Elasticsearch database population command
Addition of a command which can be launched periodically using cron, and
that handles the population of the Elasticsearch database.
This command handles every required steps of the ES database
population (including geolocation) and only requires the small change
that's been done in the model (Job table) : the Job table now has a new
column : already_in_elasticsearch.
Removal of unused code that's now replaced by the new periodic command.