Query builder skeleton
Layout of the query builder library. This library will allow flexible
replacement of the database driver responsible for the job offer queries,
without changing any code but code from the model (no need to change the
controller or the view, only the model). To swap the database driver, you'll
only have to inherit a new class from QueryTranslator which will do the
translation from an abstract query Query to the actual query which will be
executed by the driver.