Improved the synchronisation system between the two db
The populate-es command is now smarter when it comes to populating the jobs
index. Indeed, not only does it populate it, but it also use the dirty flag
from the Postgresql database to perform synchronisation operations between
the two databases, and making sure the data from the Elasticsearch database
are up to date and in sync with the data in Postgresql.