Geolocation: added a command to compute the geolocation
To perform the geocoding operation, the new command gearbox geocode must now
be used. This new command, along with the new populate-es behaviour, make it
so that the population/synchronisation of Elasticsearch and the other
operations (such as geocoding) are logically separated in their own files and
packages. They can also be run in parallel without any issues.