Bugfix: Github bot, wrong duplicate management
The job offer title isn't supposed to be used as a unique field. The only
element of a job offer which allows us to determine whether if the current job
offer is already present in the README.md file (in our case), is the job offer
url. This fixes wrong duplicates, and updates the job offer list correctly.