Changed the way the Github bot commits new offers
The Github bot now commits new job offers 1 by 1. Each commit message is now
the title of the new job offer. This version doesn't handle the case when
every element of the job offer list have to be replaced: the first commits
are a bit weird and do not really make sense.