Changed the primary key of the company table
The new primary key is now the slug of the company name.
Getting the Siren number of a company in order to submit
the form was tedious. It's much more easier for the easier
to only specify the company name. We risk having duplicates
but since every company submission is to be verified by
pyjobs's administrator before being rendered public,
this isn't really a problem (duplicates can be deleted
by the admin).