Bugfix: nominatim address url on list pages
On list pages (such as the index of both jobs and
companies controllers), there was a minor bug caused
by the encoding into utf-8 of the address while
building the link to nominatim. This would cause
unwanted behaviours when going to that link.
(Nominatim would get utf-8 characters and wouldn't
decode them, therefore, would be looking at the wrong
address).