Admin interface: "validation" changes
First draft of a solution (implementation changes will be
made, it's not definitive yet) which checks whether there
have been changes in a row while editing it (/jobs/id/edit
for instance).
If it's the case then actions are if there have or haven't
had any changes to the row.
This allows the possibility to check if the address field
of a job offer (or a company) has been modified, and it's
the case, then we can reset the geolocation_is_valid field
of the corresponding row, to make sure the new geolocation
will be recomputed.