Elasticsearch configuration
You can now specify a project name and an instance name for
Elasticsearch in the .ini configuration file.
These two parameters will be used as a prefix to every indexes
created by pyjobs like so:

        projectName_projectInstance_pyjobsIndex

Therefore, you can have multiple instances of pyjobs running on
the same Elasticsearch service, making it possible to have a
production version sitting next to the development version, or
whatsoever.