Pyjobs' processes locking
Added some support for interprocesses locking, ensuring that
multiple instance of the same process cannot perform simultaneous
identical operations where that might be dangerous. It's mostly to
make sure that if one of these processes takes to long to finish,
and cron starts them again, then the processes don't shred the
databases content to pieces.