Geocompletion: better results
If a town is spreaded on several postal codes, the previous implementation
would sometime return one that's not the smallest of all the available
postal codes.
For instance for Gre it would first return 38000 but for Grenoble 38100.
The results are now sorted on postalcodes before extracting the top_hits
for the name aggregation, giving consistent values.
The other instances of the city with different postal codes are still
accessible by providing said postal code (with or without) the city name.